Feature/client

+++ src/main/java/ar/com/ada/maven/controller/CustomerController.java
@@ -167,9 +167,11 @@ public class CustomerController {
 
     private static void deleteSelectedCustomer(int id) {
         CustomerDTO customer = customerDAO.findById(id);
-        if (customer != null) {
-            Integer toDelete = view.getResponseToDelete(customer);
-            if (toDelete == 1) {
+
+        if (customer != null && customer.getAccounts().isEmpty()) {
+            Boolean toDelete = view.getResponseToDelete(customer);
+            if (toDelete) {
+
 
                 Boolean isDelete = customerDAO.delete(id);
